Refactor `builtin-features.cpp` into multiple files
Before this commit all built-in features were kept not very well
organized inside the `builtin-features.cpp` file.

Now 4 extra files have been created:

1. `builtin-features/functions.h`
2. `builtin-features/operations.h`
3. `builtin-features/reservedWords.h`
4. `builtin-features/typeSpecificFunctions.h`

The `builtin-features.cpp` file still exists but now it only
contains `#includes "..."` of the other 4 files.

This was done to make it easier to understand each file and
easier to find an specific function if required.