Add a special meaning to attribute __type__ of maps
If the user wanted to use a map to define a custom new class
the `default_type()` would always return "map" as the type name.

Now it is possible for the user to add its own type name, e.g.:

```C++
my_type_TokenMap["__type__"] = "my_type_name";
```

This adds a special meaning to any attributes named __type__, but
it is removable if the user prefers, since the code of the
default_type function is on `builtin-function/functions.h`.