Replace `pack` for `std::shared_ptr`
The reference counting system implemented on `pack` was the same
implemented on ´std::shared_ptr` and with no bugs.

To make the replacement keep the good aspects of the pack superclass
a new superclass named `Container` was added, it has a single
shared pointer as attribute and defines a default contructor and
some few operator overloads.