Remove `packToken::operator TokenBase*()` for safety
This operator caused implicit conversions to TokenBase* on
unexpected situations.

A specially problematic one was when comparing a packToken with
a boolean. Old compilers instead of converting the boolean to
packToken would also consider converting the packToken and the
boolean to `TokenBase*`, which was crazy but anyway.

This bug was problematic because it was easy to occur and would
pop up only when using an old gcc versions (tested on gcc 4.8).