Add support for custom right unary operators
Now it is possible to design unary operators to work on the
right side of an operand, e.g. `a~` instead of `~a`.

Note, however, that an operator such as `a++` is not yet possible
because it alters the value of the lvalue `a` and this value
is not yet available for access by custom operators.

This commit also adds bug fixes and new tests for left unary operators.