Fix kwargs feature: now it replaces positional args
Before this commit you could pass keyword arguments to a function
and it would be accessible inside a variable `kwargs` when the
function was executed.

However, if the keyword argument had the same name as a positional
argument it should instead replace the argument as in Python.

It is now fixed and it is possible to set positional arguments
by name as it was originally designed to be, e.g.:

```C++
// Passing arguments to pow by name:
packToken result = calculator::calculate("pow('exp': 3, 'number': 2)");
std::cout << result << std::endl; // 8

// Works independently from arguments order:
packToken result = calculator::calculate("pow('exp': 3, 'number': 2)");
std::cout << result << std::endl; // 8
```