Move hard-coded parsing of "." operator to custom parser
For the "." operator to work it was necessary to convert the token
that came just after it to a string type instead of a variable type.

I.e. `{a: 10} . a` needs to be equivalent to `{a: 10} . 'a'`

This was being done inside the calculator class with
a hard-coded `if`.

Now with the custom parsers available it is not necessary anymore
so it was moved to `builtin-features/reservedWords.h`.