Operation parsing will now ignore '[' and '{' at the end
Before this commit the operator parser would get the greatest
string containing non-alphanumeric characters unless one of them
was in the list: `+-'"()_`

This commit adds "[]" and "{}" to this list making possible
for operations such as:

```C++
calculator::calculate("['item1']==['item1']");
```

To work. Before it would interpret `==[` as a single operator
instead of reading 2 operators.