Explicitly include `cmath` in functions.h and operations.h
The files:

- builtin-features/functions.h
- builtin-features/operations.h

Would cause an error if they were copied to be customized by a
user of the framework because they depended on the cmath lib
that is not included by the `shunting-yard.h` file.

To make it easier to the user this lib is now #included on both
files.