Improve builtin `packToken::str_custom()` implementation
Before this commit the packToken::str() function when working
on a MAP would check if the map had a "__str__" attribute,
and if yes it would use it to stringify the MAP.

However, there was no way to do a simillar specialization for
other types such as user defined types.

Now if you add a type specific function for a user defined type
named "__str__" packToken::str() will use it to stringify any
instances of this type.