Make LValues accessible for custom operations
Before this commit it was not possible to create operations
that altered the original value of their operands such as:

- `a = b`
- `a++`
- `a *= 5`

The only one that existed was the assignment operator "=" and it
was hardcoded so no one could alter its behavior.

Now it is possible not only to add new operations of that type
but also to alter the behavior of the default assignemnt "="
operator.